# src/storage/api-docs/api-docs-v2.yaml
...
openapi: 3.0.0
info:
  title: Customer Portal API
  description: Customer Portal API
  contact:
    email: alex.sviachko@goaptive.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 3.5.0
servers:
  # Added by API Auto Mocking Plugin
  - url: https://virtserver.swaggerhub.com/Aptive-Environmental/Customer_Portal_API3.5.0
    description: SwaggerHub API Auto Mocking
  - url: https://api.customer-portal.prd.goaptive.com/api/v2
    description: Production server (uses live data)
  - url: https://api.customer-portal.tst.goaptive.com/api/v2
    description: Sandbox server (uses test data)
  - url: https://api.customer-portal.stg.goaptive.com/api/v2
    description: Stage server (uses test data)
  - url: http://localhost:8080/api/v2
    description: Dev API
...
paths:
...
  /user/accounts:
    get:
      security:
        - bearerAuth: []
      tags:
        - user
      summary: List accounts for user
      description: List accounts for user currently logged in
      operationId: user.accounts
      responses:
        "200":
          description: Accounts list retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PestRoutesAccount'
        "403":
          $ref: '#/components/responses/403'
        "412":
          $ref: '#/components/responses/412'
        "422":
          $ref: '#/components/responses/InvalidRequestBodyError'
        "5XX":
          $ref: '#/components/responses/5XX'
  /customer/accounts:
    get:
      security:
        - bearerAuth: []
      tags:
        - customer
      summary: List accounts for user
      description: List accounts for user currently logged in
      operationId: customer.accounts
      responses:
        "200":
          description: Accounts list retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PestRoutesAccount'
        "403":
          $ref: '#/components/responses/403'
        "412":
          $ref: '#/components/responses/412'
        "422":
          $ref: '#/components/responses/InvalidRequestBodyError'
        "5XX":
          $ref: '#/components/responses/5XX'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    PestRoutesAccount:
      type: object
      properties:
        links:
          $ref: '#/components/schemas/PestRoutesAccount_links'
        data:
          $ref: '#/components/schemas/PestRoutesAccount_data'
    PestRoutesAccount_links:
      type: object
      properties:
        self:
          type: string
          example: /api/v2/user/accounts
    PestRoutesAccount_data:
      type: object
      properties:
        type:
          type: string
          example: Customer
        id:
          type: integer
          example: 93939421
        attributes:
          $ref: '#/components/schemas/PestRoutesAccount_data_attributes'
    PestRoutesAddress:
      type: object
      properties:
        countryCode:
          type: string
          example: US
        address:
          type: string
          nullable: true
          example: 1 Infinite Loop
        city:
          type: string
          nullable: true
          example: Cupertino
        stateCode:
          type: string
          nullable: true
          example: CA
        zip:
          type: string
          nullable: true
          example: "19804"
    PestRoutesPhone:
      type: object
      properties:
        phone:
          type: string
        extension:
          type: string
        isPrimary:
          type: boolean
...
    ErrorMessage:
      required:
        - message
      type: object
      properties:
        message:
          type: string
          description: Error message
          example: Fatal error
    ValidationErrorMessage:
      required:
        - errors
        - message
      type: object
      properties:
        message:
          type: string
          description: Error message
          example: The name field is required. (and 2 more errors)
        errors:
          maxProperties: 12
          minProperties: 1
          type: object
          additionalProperties:
            type: array
            items:
              type: string
              example: The name field is required.
          description: Error messages
      description: Validation error list
      example: '{"message":"The billing name field is required.","errors":{"billing_name": ["The billing name field is required."],"billing_state":["The billing state must be 2 characters.","The billing state must only contain letters."]}}'
...
    PestRoutesAccount_data_attributes_billingInformation:
      type: object
      properties:
        accountId:
          type: integer
        companyName:
          type: string
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Smith
        address:
          $ref: '#/components/schemas/PestRoutesAddress'
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
    PestRoutesAccount_data_attributes_source:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
    PestRoutesAccount_data_attributes_customerFlags:
      type: object
      properties:
        customerId:
          type: integer
        flag:
          type: string
        value:
          type: string
    PestRoutesAccount_data_attributes:
      type: object
      properties:
        officeId:
          type: integer
          example: 19
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Smith
        companyName:
          type: string
          nullable: true
          example: A Company
        spouse:
          type: string
          nullable: true
        isCommercial:
          type: boolean
          example: false
        status:
          type: integer
          description: 1 = Active, 0 = Inactive
          example: 1
        email:
          type: string
          nullable: true
          example: js@example.com
        phones:
          type: array
          items:
            $ref: '#/components/schemas/PestRoutesPhone'
        address:
          $ref: '#/components/schemas/PestRoutesAddress'
        billingInformation:
          $ref: '#/components/schemas/PestRoutesAccount_data_attributes_billingInformation'
        latitude:
          type: number
          format: float
        longitude:
          type: number
          format: float
        squareFeet:
          type: integer
        addedBy:
          type: integer
        dateAdded:
          type: string
          format: date
        dateCancelled:
          type: string
          format: date
        dateUpdated:
          type: string
          format: date
        source:
          $ref: '#/components/schemas/PestRoutesAccount_data_attributes_source'
        autoPay:
          type: string
        preferredTechId:
          type: integer
          nullable: true
        paidInAdvance:
          type: boolean
        subscriptionIds:
          type: array
          items:
            type: integer
        balance:
          type: number
          format: float
        balanceAge:
          type: integer
        responsibleBalance:
          type: number
          format: float
        responsibleBalanceAge:
          type: integer
        customerLink:
          type: string
          nullable: true
        masterAccountId:
          type: integer
          nullable: true
        preferredDayForBilling:
          type: integer
        paymentHoldDate:
          type: string
          format: date
          nullable: true
        mostRecentCardLastFour:
          type: string
        mostRecentCardExpirationDate:
          type: string
        regionId:
          type: integer
        mapCode:
          type: string
          nullable: true
        mapPage:
          type: string
          nullable: true
        specialScheduling:
          type: string
          nullable: true
        taxRate:
          type: number
          format: float
        smsReminders:
          type: boolean
        phoneReminders:
          type: boolean
        emailReminders:
          type: boolean
        customerSource:
          $ref: '#/components/schemas/PestRoutesAccount_data_attributes_source'
        maxMonthlyCharge:
          type: number
          format: float
        county:
          type: string
          nullable: true
        useStructures:
          type: boolean
        isMultiUnit:
          type: boolean
        autoPayPaymentProfileId:
          type: integer
          nullable: true
        divisionId:
          type: integer
          nullable: true
        appointmentIds:
          type: array
          items:
            type: integer
        ticketIds:
          type: array
          items:
            type: integer
        paymentIds:
          type: array
          items:
            type: integer
        unitIds:
          type: array
          items:
            type: integer
        customerFlags:
          type: array
          items:
            $ref: '#/components/schemas/PestRoutesAccount_data_attributes_customerFlags'
        portalLogin:
          type: string
          nullable: true
        portalLoginExpires:
          type: string
          format: date
          nullable: true
...
  responses:
...
    "403":
      description: User is not authenticated
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'
...
    "412":
      description: User cannot be fetched from PestRoutes API
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'
    "5XX":
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'
    "InvalidRequestBodyError":
      description: Invalid request body error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorMessage'
...